+++ rust/operator-binary/src/controller.rs
@@ -1,19 +1,23 @@
 //! Ensures that `Pod`s are configured and running for each [`TrinoCluster`]
 use snafu::{OptionExt, ResultExt, Snafu};
-use stackable_operator::builder::{PodSecurityContextBuilder, VolumeBuilder};
-use stackable_operator::commons::s3::{S3AccessStyle, S3ConnectionDef};
-use stackable_operator::kube::runtime::reflector::ObjectRef;
 use stackable_operator::{
-    builder::{ConfigMapBuilder, ContainerBuilder, ObjectMetaBuilder, PodBuilder},
+    builder::{
+        ConfigMapBuilder, ContainerBuilder, ObjectMetaBuilder, PodBuilder,
+        PodSecurityContextBuilder, VolumeBuilder,
+    },
     client::Client,
-    commons::{opa::OpaApiVersion, s3::S3ConnectionSpec},
+    commons::{
+        opa::OpaApiVersion,
+        s3::{S3AccessStyle, S3ConnectionDef, S3ConnectionSpec},
+    },
     k8s_openapi::{
         api::{
             apps::v1::{StatefulSet, StatefulSetSpec},
             core::v1::{
                 CSIVolumeSource, ConfigMap, ConfigMapVolumeSource, ContainerPort, EnvVar,
-                PersistentVolumeClaim, PersistentVolumeClaimSpec, Probe, ResourceRequirements,
-                SecurityContext, Service, ServicePort, ServiceSpec, TCPSocketAction, Volume,
+                EnvVarSource, PersistentVolumeClaim, PersistentVolumeClaimSpec, Probe,
+                ResourceRequirements, Secret, SecretKeySelector, SecurityContext, Service,
+                ServicePort, ServiceSpec, TCPSocketAction, Volume,
             },
         },
         apimachinery::pkg::{
@@ -22,7 +26,7 @@ use stackable_operator::{
     },
     kube::{
         api::ObjectMeta,
-        runtime::controller::{Action, Context},
+        runtime::{controller::Action, reflector::ObjectRef},
         ResourceExt,
     },
     labels::{role_group_selector_labels, role_selector_labels},
@@ -40,16 +44,18 @@ use stackable_trino_crd::{
     authentication,
     authentication::TrinoAuthenticationConfig,
     discovery::{TrinoDiscovery, TrinoDiscoveryProtocol, TrinoPodRef},
-    TrinoCluster, TrinoClusterSpec, TrinoRole, ACCESS_CONTROL_PROPERTIES, APP_NAME,
-    CONFIG_DIR_NAME, CONFIG_PROPERTIES, DATA_DIR_NAME, DISCOVERY_URI, HIVE_PROPERTIES, HTTPS_PORT,
-    HTTPS_PORT_NAME, HTTP_PORT, HTTP_PORT_NAME, JVM_CONFIG, KEYSTORE_DIR_NAME, LOG_PROPERTIES,
-    METRICS_PORT, METRICS_PORT_NAME, NODE_PROPERTIES, PASSWORD_AUTHENTICATOR_PROPERTIES,
-    PASSWORD_DB, RW_CONFIG_DIR_NAME, S3_ACCESS_KEY, S3_ENDPOINT, S3_PATH_STYLE_ACCESS,
-    S3_SECRET_DIR_NAME, S3_SECRET_KEY, S3_SSL_ENABLED, USER_PASSWORD_DATA_DIR_NAME,
+    TrinoCluster, TrinoRole, ACCESS_CONTROL_PROPERTIES, APP_NAME, CONFIG_DIR_NAME,
+    CONFIG_PROPERTIES, DATA_DIR_NAME, DISCOVERY_URI, HIVE_PROPERTIES, HTTPS_PORT, HTTPS_PORT_NAME,
+    HTTP_PORT, HTTP_PORT_NAME, INTERNAL_COMMUNICATION_SHARED_SECRET, JVM_CONFIG, KEYSTORE_DIR_NAME,
+    LOG_PROPERTIES, METRICS_PORT, METRICS_PORT_NAME, NODE_PROPERTIES,
+    PASSWORD_AUTHENTICATOR_PROPERTIES, PASSWORD_DB, RW_CONFIG_DIR_NAME, S3_ACCESS_KEY, S3_ENDPOINT,
+    S3_PATH_STYLE_ACCESS, S3_SECRET_DIR_NAME, S3_SECRET_KEY, S3_SSL_ENABLED,
+    USER_PASSWORD_DATA_DIR_NAME,
 };
 use std::borrow::Cow;
 use std::{
     collections::{BTreeMap, HashMap},
+    fmt::Write,
     str::FromStr,
     sync::Arc,
     time::Duration,
@@ -69,8 +75,6 @@ const FIELD_MANAGER_SCOPE: &str = "trinocluster";
 #[strum_discriminants(derive(IntoStaticStr))]
 #[allow(clippy::enum_variant_names)]
 pub enum Error {
-    #[snafu(display("object defines no version"))]
-    ObjectHasNoVersion,
     #[snafu(display("object defines no namespace"))]
     ObjectHasNoNamespace,
     #[snafu(display("object defines no {} role", role))]
@@ -110,6 +114,10 @@ pub enum Error {
         source: stackable_operator::error::Error,
         rolegroup: RoleGroupRef<TrinoCluster>,
     },
+    #[snafu(display("failed to apply internal secret"))]
+    ApplyInternalSecret {
+        source: stackable_operator::error::Error,
+    },
     #[snafu(display("invalid product config"))]
     InvalidProductConfig {
         source: stackable_operator::error::Error,
@@ -144,6 +152,8 @@ pub enum Error {
     },
     #[snafu(display("invalid S3 connection: {reason}"))]
     InvalidS3Connection { reason: String },
+    #[snafu(display("failed to parse trino product version"))]
+    TrinoProductVersionParseFailure { source: stackable_trino_crd::Error },
 }
 
 type Result<T, E = Error> = std::result::Result<T, E>;
@@ -158,12 +168,15 @@ const ENV_S3_ACCESS_KEY: &str = "S3_ACCESS_KEY";
 const ENV_S3_SECRET_KEY: &str = "S3_SECRET_KEY";
 const SECRET_KEY_S3_ACCESS_KEY: &str = "accessKey";
 const SECRET_KEY_S3_SECRET_KEY: &str = "secretKey";
+const INTERNAL_SECRET: &str = "INTERNAL_SECRET";
 
-pub async fn reconcile_trino(trino: Arc<TrinoCluster>, ctx: Context<Ctx>) -> Result<Action> {
+pub async fn reconcile_trino(trino: Arc<TrinoCluster>, ctx: Arc<Ctx>) -> Result<Action> {
     tracing::info!("Starting reconcile");
 
-    let client = &ctx.get_ref().client;
-    let version = trino_version(&trino)?;
+    let client = &ctx.client;
+    let trino_product_version = trino
+        .product_version()
+        .context(TrinoProductVersionParseFailureSnafu)?;
 
     let catalogs = client
         .list_with_label_selector::<TrinoCatalog>(
@@ -186,7 +199,7 @@ pub async fn reconcile_trino(trino: Arc<TrinoCluster>, ctx: Context<Ctx>) -> Res
         .collect::<Result<Vec<_>>>()?;
 
     let mut validated_config =
-        validated_product_config(&trino, version, &ctx.get_ref().product_config)?;
+        validated_product_config(&trino, &trino_product_version, &ctx.product_config)?;
 
     let s3_connection_def: &Option<S3ConnectionDef> = &trino.spec.s3;
     let s3_connection_spec: Option<S3ConnectionSpec> = if let Some(s3) = s3_connection_def {
@@ -273,6 +286,8 @@ pub async fn reconcile_trino(trino: Arc<TrinoCluster>, ctx: Context<Ctx>) -> Res
         .await
         .context(ApplyRoleServiceSnafu)?;
 
+    create_shared_internal_secret(&trino, client).await?;
+
     for (role, role_config) in validated_config {
         let trino_role = TrinoRole::from_str(&role).context(InternalOperatorFailureSnafu)?;
         for (role_group, config) in role_config {
@@ -347,7 +362,15 @@ pub fn build_coordinator_role_service(trino: &TrinoCluster) -> Result<Service> {
             .name(&role_svc_name)
             .ownerreference_from_resource(trino, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(trino, APP_NAME, trino_version(trino)?, &role_name, "global")
+            .with_recommended_labels(
+                trino,
+                APP_NAME,
+                trino
+                    .image_version()
+                    .context(TrinoProductVersionParseFailureSnafu)?,
+                &role_name,
+                "global",
+            )
             .build(),
         spec: Some(ServiceSpec {
             ports: Some(service_ports(trino)),
@@ -409,6 +432,12 @@ fn build_rolegroup_config_map(
                     Some(discovery.connection_string()),
                 );
 
+                // Required from Trino 378 (accepted in 377)
+                transformed_config.insert(
+                    INTERNAL_COMMUNICATION_SHARED_SECRET.to_string(),
+                    Some(format!("${{ENV:{secret}}}", secret = INTERNAL_SECRET)),
+                );
+
                 let config_properties =
                     product_config::writer::to_java_properties_string(transformed_config.iter())
                         .context(FailedToWriteJavaPropertiesSnafu)?;
@@ -441,7 +470,7 @@ fn build_rolegroup_config_map(
                 cm_conf_data.insert(file_name.to_string(), "".to_string());
             }
             PropertyNameKind::File(file_name) if file_name == JVM_CONFIG => {
-                jvm_config.push_str(&format!("\n-javaagent:/stackable/jmx/jmx_prometheus_javaagent-0.16.1.jar={}:/stackable/jmx/config.yaml", METRICS_PORT));
+                let _ = write!(jvm_config, "\n-javaagent:/stackable/jmx/jmx_prometheus_javaagent-0.16.1.jar={}:/stackable/jmx/config.yaml", METRICS_PORT);
             }
             _ => {}
         }
@@ -474,7 +503,9 @@ fn build_rolegroup_config_map(
                 .with_recommended_labels(
                     trino,
                     APP_NAME,
-                    trino_version(trino)?,
+                    trino
+                        .image_version()
+                        .context(TrinoProductVersionParseFailureSnafu)?,
                     &rolegroup_ref.role,
                     &rolegroup_ref.role_group,
                 )
@@ -504,7 +535,9 @@ fn build_rolegroup_catalog_config_map(
                 .with_recommended_labels(
                     trino,
                     APP_NAME,
-                    trino_version(trino)?,
+                    trino
+                        .image_version()
+                        .context(TrinoProductVersionParseFailureSnafu)?,
                     &rolegroup_ref.role,
                     &rolegroup_ref.role_group,
                 )
@@ -558,7 +591,9 @@ fn build_rolegroup_statefulset(
         })?
         .role_groups
         .get(&rolegroup_ref.role_group);
-    let trino_version = trino_version_trim(trino)?;
+    let trino_image_version = trino
+        .image_version()
+        .context(TrinoProductVersionParseFailureSnafu)?;
 
     let mut env = config
         .get(&PropertyNameKind::Env)
@@ -581,6 +616,11 @@ fn build_rolegroup_statefulset(
         value: Some("hadoop-azure-datalake".to_string()),
         value_from: None,
     });
+
+    let secret_name = build_shared_internal_secret_name(trino);
+    if let Some(internal_secret) = env_var_from_secret(&Some(secret_name), INTERNAL_SECRET) {
+        env.push(internal_secret);
+    };
 
     // Add volume and volume mounts for s3 credentials
     if let Some(S3ConnectionSpec {
@@ -609,8 +649,8 @@ fn build_rolegroup_statefulset(
 
     let container_trino = container_builder
         .image(format!(
-            "docker.stackable.tech/stackable/trino:{}-stackable0",
-            trino_version
+            "docker.stackable.tech/stackable/trino:{}",
+            trino_image_version
         ))
         .command(vec!["/bin/bash".to_string(), "-c".to_string()])
         .args(container_trino_args(authentication_config, s3_connection))
@@ -634,7 +674,7 @@ fn build_rolegroup_statefulset(
             .with_recommended_labels(
                 trino,
                 APP_NAME,
-                trino_version,
+                trino_image_version,
                 &rolegroup_ref.role,
                 &rolegroup_ref.role_group,
             )
@@ -661,7 +701,7 @@ fn build_rolegroup_statefulset(
                     m.with_recommended_labels(
                         trino,
                         APP_NAME,
-                        trino_version,
+                        trino_image_version,
                         &rolegroup_ref.role,
                         &rolegroup_ref.role_group,
                     )
@@ -746,7 +786,9 @@ fn build_rolegroup_service(
             .with_recommended_labels(
                 trino,
                 APP_NAME,
-                trino_version(trino)?,
+                trino
+                    .image_version()
+                    .context(TrinoProductVersionParseFailureSnafu)?,
                 &rolegroup.role,
                 &rolegroup.role_group,
             )
@@ -768,25 +810,7 @@ fn build_rolegroup_service(
     })
 }
 
-/// Returns our semver representation for product config e.g. 0.0.377
-pub fn trino_version(trino: &TrinoCluster) -> Result<&str> {
-    trino
-        .spec
-        .version
-        .as_deref()
-        .context(ObjectHasNoVersionSnafu)
-}
-
-/// Returns the "real" Trino version for docker images e.g. 377
-pub fn trino_version_trim(trino: &TrinoCluster) -> Result<&str> {
-    let spec: &TrinoClusterSpec = &trino.spec;
-    spec.version
-        .as_deref()
-        .and_then(|v| v.split('.').collect::<Vec<_>>().last().cloned())
-        .context(ObjectHasNoVersionSnafu)
-}
-
-pub fn error_policy(_error: &Error, _ctx: Context<Ctx>) -> Action {
+pub fn error_policy(_error: &Error, _ctx: Arc<Ctx>) -> Action {
     Action::requeue(Duration::from_secs(5))
 }
 
@@ -819,12 +843,12 @@ fn container_prepare_args() -> Vec<String> {
         "echo Cleaning up truststore - just in case",
         &format!("rm -f {keystore_directory}/truststore.p12", keystore_directory = KEYSTORE_DIR_NAME),
         "echo Creating truststore",
-        &format!("keytool -importcert -file {keystore_directory}/ca.crt -keystore {keystore_directory}/truststore.p12 -storetype pkcs12 -noprompt -alias ca_cert -storepass secret", 
+        &format!("keytool -importcert -file {keystore_directory}/ca.crt -keystore {keystore_directory}/truststore.p12 -storetype pkcs12 -noprompt -alias ca_cert -storepass secret",
                  keystore_directory = KEYSTORE_DIR_NAME),
         "echo Creating certificate chain",
         &format!("cat {keystore_directory}/ca.crt {keystore_directory}/tls.crt > {keystore_directory}/chain.crt", keystore_directory = KEYSTORE_DIR_NAME),
         "echo Creating keystore",
-        &format!("openssl pkcs12 -export -in {keystore_directory}/chain.crt -inkey {keystore_directory}/tls.key -out {keystore_directory}/keystore.p12 --passout file:{keystore_directory}/password", 
+        &format!("openssl pkcs12 -export -in {keystore_directory}/chain.crt -inkey {keystore_directory}/tls.key -out {keystore_directory}/keystore.p12 --passout file:{keystore_directory}/password",
                  keystore_directory = KEYSTORE_DIR_NAME),
         "echo Cleaning up password",
         &format!("rm -f {keystore_directory}/password", keystore_directory = KEYSTORE_DIR_NAME),
@@ -916,6 +940,21 @@ fn container_trino_args(
     vec![args.join(" && ")]
 }
 
+fn env_var_from_secret(secret_name: &Option<String>, env_var: &str) -> Option<EnvVar> {
+    secret_name.as_ref().map(|secret| EnvVar {
+        name: env_var.to_string(),
+        value_from: Some(EnvVarSource {
+            secret_key_ref: Some(SecretKeySelector {
+                optional: Some(false),
+                name: Some(secret.to_string()),
+                key: env_var.to_string(),
+            }),
+            ..EnvVarSource::default()
+        }),
+        ..EnvVar::default()
+    })
+}
+
 /// Defines all required roles and their required configuration.
 ///
 /// The roles and their configs are then validated and complemented by the product config.
@@ -979,6 +1018,49 @@ fn validated_product_config(
 
     validate_all_roles_and_groups_config(version, &role_config, product_config, false, false)
         .context(InvalidProductConfigSnafu)
+}
+
+async fn create_shared_internal_secret(trino: &TrinoCluster, client: &Client) -> Result<()> {
+    let secret = build_shared_internal_secret(trino)?;
+    if !client
+        .exists::<Secret>(&secret.name(), secret.namespace().as_deref())
+        .await
+        .context(ApplyInternalSecretSnafu)?
+    {
+        client
+            .apply_patch(FIELD_MANAGER_SCOPE, &secret, &secret)
+            .await
+            .context(ApplyInternalSecretSnafu)?;
+    }
+
+    Ok(())
+}
+
+fn build_shared_internal_secret(trino: &TrinoCluster) -> Result<Secret> {
+    let mut internal_secret = BTreeMap::new();
+    internal_secret.insert(INTERNAL_SECRET.to_string(), get_random_base64());
+
+    Ok(Secret {
+        immutable: Some(true),
+        metadata: ObjectMetaBuilder::new()
+            .name(build_shared_internal_secret_name(trino))
+            .namespace_opt(trino.namespace())
+            .ownerreference_from_resource(trino, None, Some(true))
+            .context(ObjectMissingMetadataForOwnerRefSnafu)?
+            .build(),
+        string_data: Some(internal_secret),
+        ..Secret::default()
+    })
+}
+
+fn build_shared_internal_secret_name(trino: &TrinoCluster) -> String {
+    format!("{}-internal-secret", trino.name())
+}
+
+fn get_random_base64() -> String {
+    let mut buf = [0; 512];
+    openssl::rand::rand_bytes(&mut buf).unwrap();
+    openssl::base64::encode_block(&buf)
 }
 
 fn service_ports(trino: &TrinoCluster) -> Vec<ServicePort> {